requirements.txt
fastapi==0.95.2
uvicorn==0.15.0
gunicorn==20.1.0
python-dotenv==0.19.0
google-generativeai==0.3.0
pillow==8.3.1
numpy==1.21.2
pandas==1.3.3
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
pydantic==1.10.7
pydantic-settings==2.0.3 
grpcio==1.59.3